merge upstream

+++ services/admin/UrlKey.php
@@ -20,9 +20,19 @@ use fecshop\services\Service;
  */
 class UrlKey extends Service
 {
+    
+    const URLKEY_LABEL_ARR = 'appadmin_urlkey_label_cache_arr'; 
+    
     public $numPerPage = 20;
 
     public $urlKeyTags;
+    protected $_urlKeyTags;
+    
+    // 没有在数据库中添加的url key
+    public $addUrlKeyAndLabelArr = [
+        '/fecadmin/login/index' => '帐号登陆',
+        '/fecadmin/logout/index' => '帐号退出',
+    ];
 
     protected $_staticBlockModelName = '\fecshop\models\mysqldb\admin\UrlKey';
 
@@ -41,7 +51,14 @@ class UrlKey extends Service
     }
 
     public function getTags(){
-        return $this->urlKeyTags;
+        if (!$this->_urlKeyTags) {
+            if (is_array($this->urlKeyTags)) {
+                foreach ($this->urlKeyTags as $k => $v) {
+                    $this->_urlKeyTags[$k] = Yii::$service->page->translate->__($v);
+                }
+            }
+        }
+        return $this->_urlKeyTags;
     }
 
     public function getPrimaryKey()
@@ -97,6 +114,44 @@ class UrlKey extends Service
             'count'=> $query->limit(null)->offset(null)->count(),
         ];
     }
+    /**
+     * @return array 得到urlKey 和 label 对应的数组。
+     *  这样可以通过url_key得到当前操作的菜单的name
+     */
+    public function getUrlKeyAndLabelArr(){
+        $arr = Yii::$app->cache->get(self::URLKEY_LABEL_ARR);
+        if (!$arr) {
+            $arr = [];
+            $filter = [
+                'fetchAll' => true,
+                'asArray' => true,
+            ];
+            $data = $this->coll($filter);
+            if (is_array($data['coll'])) {
+                $tags = $this->getTags();
+                foreach ($data['coll'] as $one) {
+                    $url_key =  $one['url_key'];
+                    $label = $tags[$one['tag']] .' '. $one['name'];
+                    $arr[$url_key] = $label;
+                }
+            }
+            $addArr = $this->getAddUrlKeyAndLabelArr();
+            if (is_array($addArr)) {
+                $arr = array_merge($arr, $addArr);
+            }
+            Yii::$app->cache->set(self::URLKEY_LABEL_ARR, $arr);
+        }
+        
+        return $arr;
+    }
+    
+    protected function getAddUrlKeyAndLabelArr(){
+        
+        return $this->addUrlKeyAndLabelArr;
+    }
+    
+    
+    
 	/**
 	 * @return 按照tag，将资源（url_key）进行分组, 按照 tag_sort_order 进行排序
      * 根据传递的roleId 获取对应的url_key_id，将资源（url_key）默认选中
@@ -105,12 +160,11 @@ class UrlKey extends Service
         $selectRoleUrlKeys = [];
         if ($role_id) {
             $filter = [
-                'numPerPage' 	=> 9000,
-                'pageNum'		=> 1,
                 'where'			=> [
                     ['role_id' => $role_id],
                 ],
                 'asArray' => true,
+                'fetchAll' => true,
             ];
             $data = Yii::$service->admin->roleUrlKey->coll($filter);
             if (isset($data['coll']) && is_array($data['coll'])) {
@@ -121,8 +175,7 @@ class UrlKey extends Service
         }
 		$filter = [ 
 			'asArray' => true,
-			'numPerPage' 	=> 4000,
-      		'pageNum'		=> 1,
+			'fetchAll' => true,
 		];
 		$coll = $this->coll($filter);
 		$arr = [];
@@ -144,8 +197,12 @@ class UrlKey extends Service
                 $arr[$tag][] = $one_arr;
 			}
 		}
-		$arr = \fec\helpers\CFunc::array_sort($arr, 'tag_sort_order', 'asc', true);
-        $urlKeyTags = $this->urlKeyTags;
+        // 按照 tag_sort_order 进行排序
+        foreach ($arr as $k => $one) {
+            $arr[$k] = \fec\helpers\CFunc::array_sort($one, 'tag_sort_order', 'asc', true);
+        }
+        //var_dump($arr);
+        $urlKeyTags = $this->getTags();
         $arrSort = [];
         foreach ($urlKeyTags as $k => $v) {
             if (isset($arr[$k])) {
@@ -156,6 +213,7 @@ class UrlKey extends Service
 		return $arrSort;
 	}
 
+    public $can_not_delete = 1;
     /**
      * @property $one|array
      * save $data to cms model,then,add url rewrite info to system service urlrewrite.
@@ -174,7 +232,12 @@ class UrlKey extends Service
             if (!$model) {
                 Yii::$service->helper->errors->add('static block '.$this->getPrimaryKey().' is not exist');
 
-                return;
+                return false;
+            }
+            if ($model->can_delete == $this->can_not_delete) {
+                Yii::$service->helper->errors->add('resource(url key) created by system, can not edit and save');
+
+                return false;
             }
         } else {
             $model = new $this->_staticBlockModelName();
@@ -222,14 +285,27 @@ class UrlKey extends Service
         if (is_array($ids) && !empty($ids)) {
             foreach ($ids as $id) {
                 $model = $this->_staticBlockModel->findOne($id);
+                if ($model->can_delete == $this->can_not_delete) {
+                    Yii::$service->helper->errors->add('resource(url key) created by system, can not remove');
+
+                    return false;
+                }
+
                 $model->delete();
+                // delete roleUrlKey
+                Yii::$service->admin->roleUrlKey->removeByUrlKeyId($id);
             }
         } else {
             $id = $ids;
-            foreach ($ids as $id) {
-                $model = $this->_staticBlockModel->findOne($id);
-                $model->delete();
+            $model = $this->_staticBlockModel->findOne($id);
+            if ($model->can_delete == $this->can_not_delete) {
+                Yii::$service->helper->errors->add('resource(url key) created by system, can not remove');
+
+                return false;
             }
+            $model->delete();
+            // delete roleUrlKey
+            Yii::$service->admin->roleUrlKey->removeByUrlKeyId($id);
         }
 
         return true;
